Orca: Merge TorchRunner with TrainingOperator(train) (#6863)
* expose TrainingOperator.train

* fix horovod case
